modularised scan
