+++ zingo-sync/src/scan/task.rs
